requirements.txt
aniso8601==9.0.1
asttokens==2.2.1
attrs==23.1.0
axios==0.4.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
checksumdir==1.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dacite==1.8.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
Flask==2.3.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
flask-restful-swagger-3==0.5.1
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.39.4
funcs==0.8.0
greenlet==2.0.2
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-resources==5.12.0
ipykernel==6.23.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multimethod==1.9.1
named==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
packaging==23.1
pandas==1.5.3
pandas-profiling==3.6.6
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
rich==13.6.0
rpds-py==0.9.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
solus==1.1.0
speaklater==1.3
SQLAlchemy==2.0.3
stack-data==0.6.2
statsmodels==0.14.0
tangled-up-in-unicode==0.2.0
threadpoolctl==3.1.0
too==0.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==2.13.3
typing-aliases==1.1.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
visions==0.7.5
wcwidth==0.2.6
Werkzeug==2.3.6
wordcloud==1.9.2
wraps==0.6.1
WTForms==3.0.1
ydata-profiling==4.2.0
